# api/src/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Bon-API
  version: v0
servers:
  -
    url: 'http://localhost:8000/api/v0'
paths:
...
  /plat:
    get:
      tags:
        - Plat
      operationId: 53896f8196bd7d3e7b7bbc23b49edeac
      responses:
        '200':
          description: 'Retourne l ensemble des plats'
          content:
            application/json:
              schema:
                description: 'Contenu de notre menu'
                type: array
                items:
                  $ref: '#/components/schemas/Plat'
    post:
      tags:
        - Plat
      operationId: 8cceb4489d1a0aebc341ab344d476e97
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - prix
                - image
                - description
              properties:
                name:
                  type: string
                prix:
                  type: number
                  format: float
                image:
                  type: string
                description:
                  type: string
              type: object
              example:
                name: 'Nom du plat'
                prix: 26.2
                image: 'Image d un plat'
                description: 'Descripion du plat'
      responses:
        '201':
          description: 'Retourne le plat nouvellement crée'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plat'
      security:
        -
          Authorization: []
  '/plat/{id}':
    get:
      tags:
        - Plat
      operationId: 51c0a732a3edcebd736093134acc1ce6
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne un plat en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plat'
    put:
      tags:
        - Plat
      operationId: 4dc070f40565d9f013345fc502d12804
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - prix
                - image
                - description
              properties:
                name:
                  type: string
                prix:
                  type: number
                  format: float
                image:
                  type: string
                description:
                  type: string
              type: object
              example:
                name: 'Nom du plat'
                prix: 26.2
                image: 'Image d un plat'
                description: 'Descripion du plat'
      responses:
        '200':
          description: 'Plat modifié'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - Plat
      operationId: e20af23e8cf4ad65ab97d0636b75190e
      parameters:
        -
          name: id
          in: path
          description: 'ID du plat'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Plat supprimé'
          content:
            application/json:
              schema:
                type: string
        '404':
          description: 'Erreur de récupération'
          content:
            application/json:
              schema:
                description: 'Message d erreur'
                type: string
              example: 'Ce plat n''existe pas/plus'
      security:
        -
          Authorization: []
  /reservation:
    get:
      tags:
        - Reservation
      operationId: 715656f42fb681ff390ad458e5ba03aa
      responses:
        '200':
          description: 'Retourne l ensemble des reservations'
          content:
            application/json:
              schema:
                description: 'Toutes les reservations'
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
    post:
      tags:
        - Reservation
      operationId: 57598c5a05524c7781e54ce32c9db67e
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nom
                - nombre
                - creneaux
              properties:
                nom:
                  type: string
                nombre:
                  type: integer
                creneaux:
                  type: string
              type: object
              example:
                nom: 'John Doe'
                creneaux: '2022-03-04 19:45'
                nombre: 5
      responses:
        '200':
          description: 'Réservation créée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
  '/reservation/{id}':
    get:
      tags:
        - Reservation
      operationId: 2074a5451bec35ccc0c327e69a7516d3
      parameters:
        -
          name: id
          in: path
          description: 'ID de la reservation'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Retourne une reservation en fonction de son ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
    put:
      tags:
        - Reservation
      operationId: a5c54898b468828c95b19b182a08dd53
      parameters:
        -
          name: id
          in: path
          description: 'ID de la reservation à modifier'
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              required:
                - nom
                - nombre
                - creneaux
              properties:
                nom:
                  type: string
                nombre:
                  type: integer
                creneaux:
                  type: string
              type: object
              example:
                nom: 'John Doe'
                creneaux: '2022-03-04 19:45'
                nombre: 5
      responses:
        '200':
          description: 'Réservation modifiée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
    delete:
      tags:
        - Reservation
      operationId: f0fff1cd653ee7da2515b91905aef699
      parameters:
        -
          name: id
          in: path
          description: 'ID de la reservation'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Réservation supprimée'
          content:
            application/json:
              schema:
                type: string
      security:
        -
          Authorization: []
...
components:
  schemas:
...
    Plat:
      properties:
        id:
          type: integer
          nullable: false
        nom:
          type: string
          nullable: false
        prix:
          type: number
          format: float
          nullable: false
        image:
          type: string
          nullable: false
        description:
          type: string
          nullable: false
      type: object
...
    Reservation:
      properties:
        id:
          type: integer
          nullable: false
        nom:
          type: string
          nullable: false
        nombre:
          type: integer
          nullable: false
        creneaux:
          type: string
          nullable: false
      type: object
...
  securitySchemes:
    Authorization:
      type: http
      bearerFormat: JWT
      scheme: bearer
...